+++ changelog.md
@@ -5,6 +5,9 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Use `windows-2025` instead of `windows-2022` for Nuitka builds.
+- Add `windows-2025` to the test matrix.
+
 ## [4.14.2 (2025-02-18)](https://github.com/kdeldycke/workflows/compare/v4.14.1...v4.14.2)
 
 - Fix update of `.gitignore` with `gitignore-extra-content` content.